+++ src/main/java/fr/farmvivi/wittmer/Role.java
@@ -1,26 +1,26 @@
 package fr.farmvivi.wittmer;
 
 public enum Role {
-    ELEVE((short) 1, 753696822995714225L),
-    PROF((short) 2, 753696271872819331L),
-    DELEGUE((short) 3, 782576791197974538L);
+    ELEVE(1L, 753696822995714225L),
+    PROF(2L, 753696271872819331L),
+    DELEGUE(3L, 782576791197974538L);
 
-    private final short id;
+    private final long id;
     private final long roleId;
 
-    Role(short id, long roleId) {
+    Role(long id, long roleId) {
         this.id = id;
         this.roleId = roleId;
     }
 
-    public static Role getById(short id) {
+    public static Role getById(long id) {
         for (Role role : values())
             if (role.getId() == id)
                 return role;
         return null;
     }
 
-    public short getId() {
+    public long getId() {
         return id;
     }
